requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
matplotlib
plotly
scipy
sklearn
scikit-learn
pandas
numpy
git+https://github.com/openai/whisper.git
pytube
streamlit
streamlit_chat
openai
pinecone-client
python-dotenv